Implement comprehensive SeqSetVAE improvements for pretrain-finetune separation
