--- MODE_COMPARISON.md
@@ -1,182 +1,0 @@
-# SeqSetVAE: 全训练模式 vs 分类模式详细对比
-
-## 🔍 模式概述
-
-| 特性 | 全训练模式 (Full Training) | 分类模式 (Classification-Only) |
-|------|---------------------------|--------------------------------|
-| **使用场景** | 预训练阶段 | 微调阶段 |
-| **标志位** | `self.classification_only = False` | `self.classification_only = True` |
-| **主要目标** | 学习良好的表示 | 优化分类性能 |
-
-## 📊 核心区别详解
-
-### 1. **损失函数计算**
-
-#### 🔄 全训练模式
-```python
-# 计算完整的ELBO损失
-if self.classification_only:
-    # 不会执行这个分支
-else:
-    # 计算重建损失
-    recon_loss_total = 0.0
-    for idx, s_dict in enumerate(sets):
-        recon = self.decoder(h_seq[:, idx], N_t, noise_std=0.3)
-        recon_loss_total += chamfer_recon_loss(recon, target_x)
-    
-    # 总损失 = 预测损失 + 重建损失 + KL损失
-    total_loss = pred_weight * pred_loss + recon_weight * recon_loss + kl_loss
-```
-
-#### 🎯 分类模式
-```python
-if self.classification_only:
-    # 跳过重建，只计算分类损失
-    total_loss = pred_loss
-    recon_loss = torch.tensor(0.0)  # 设为0
-    kl_loss = torch.tensor(0.0)     # 设为0
-```
-
-### 2. **特征提取策略**
-
-#### 🔄 全训练模式 - 复杂多尺度特征提取
-```python
-if self.classification_only:
-    # 不会执行这个分支
-else:
-    # 使用复杂的多尺度pooling
-    global_avg = self.feature_fusion['global_pool'](h_t.transpose(1, 2))
-    global_max = self.feature_fusion['max_pool'](h_t.transpose(1, 2))
-    
-    # 注意力pooling
-    attn_output, _ = self.feature_fusion['attention_pool'](query, h_t, h_t)
-    
-    # 融合所有特征
-    combined_features = torch.cat([global_avg, global_max, attention_pool], dim=1)
-    enhanced_features = self.feature_projection(combined_features)
-```
-
-#### 🎯 分类模式 - 简单稳定特征提取
-```python
-if self.classification_only:
-    # 使用最新时刻的特征 + 注意力加权
-    last_token = h_t[:, -1, :]  # 最近的表示
-    
-    # 简单的注意力加权pooling
-    attn_weights = F.softmax(
-        torch.matmul(h_t, last_token.unsqueeze(-1)).squeeze(-1), dim=1
-    )
-    attn_pooled = torch.sum(h_t * attn_weights.unsqueeze(-1), dim=1)
-    
-    # 简单组合：70%最新 + 30%注意力加权
-    enhanced_features = 0.7 * last_token + 0.3 * attn_pooled
-```
-
-### 3. **VAE特征融合策略**
-
-#### 🔄 全训练模式 - 可学习门控融合
-```python
-if hasattr(self, 'vae_feature_fusion') and not self.classification_only:
-    # 复杂的可学习融合
-    mu_proj = self.vae_feature_fusion['mean_projection'](mu)
-    var_proj = self.vae_feature_fusion['var_projection'](std)
-    
-    # 学习融合权重
-    fusion_gate = self.vae_feature_fusion['fusion_gate'](
-        torch.cat([mu_proj, var_proj], dim=-1)
-    )
-    
-    # 门控组合
-    fused = fusion_gate * mu_proj + (1 - fusion_gate) * var_proj
-    
-    # 不确定性校准
-    uncertainty_score = self.vae_feature_fusion['uncertainty_calibration'](std)
-    fused = fused * (1.0 + 0.1 * uncertainty_score)
-```
-
-#### 🎯 分类模式 - 简单不确定性加权
-```python
-else:  # classification_only = True
-    # 简单但有效的融合
-    uncertainty = torch.mean(std, dim=-1, keepdim=True)
-    uncertainty_weight = torch.sigmoid(-uncertainty + 1.0)
-    
-    # 方差调制均值特征
-    variance_modulation = 1.0 + 0.05 * torch.tanh(std)
-    modulated_mu = mu * variance_modulation
-    
-    # 不确定性加权组合
-    fused = uncertainty_weight * modulated_mu + (1 - uncertainty_weight) * mu
-```
-
-### 4. **优化器配置**
-
-#### 🔄 全训练模式
-```python
-else:  # not classification_only
-    # 多组参数，不同学习率
-    setvae_params = list(self.setvae.parameters())
-    transformer_params = list(self.transformer.parameters())
-    cls_params = list(self.cls_head.parameters())
-    
-    optimizer = AdamW([
-        {'params': setvae_params, 'lr': self.lr * 0.5},
-        {'params': transformer_params, 'lr': self.lr},
-        {'params': cls_params, 'lr': self.lr * 2.0}
-    ])
-```
-
-#### 🎯 分类模式
-```python
-if self.classification_only:
-    # 只优化分类头
-    cls_params = [p for p in self.cls_head.parameters() if p.requires_grad]
-    optimizer = AdamW(
-        [{'params': cls_params, 'lr': cls_lr}],
-        weight_decay=0.01  # 更强的正则化
-    )
-```
-
-### 5. **模型状态管理**
-
-#### 🔄 全训练模式
-- 所有模块都在训练模式
-- 允许dropout、batch norm等随机性
-- 参数更新影响整个网络
-
-#### 🎯 分类模式
-```python
-def set_backbone_eval(self):
-    # 强制backbone为eval模式
-    self.setvae.eval()
-    self.transformer.eval()
-    self.post_transformer_norm.eval()
-    self.decoder.eval()
-    self.feature_fusion.eval()
-    self.vae_feature_fusion.eval()
-
-def on_train_start(self):
-    if self.classification_only:
-        self.set_backbone_eval()  # 每次训练开始时强制eval
-```
-
-## 🎯 为什么需要两种模式？
-
-### 全训练模式的优势
-1. **表示学习**: 通过重建任务学习有意义的潜在表示
-2. **正则化**: KL损失防止后验坍塌
-3. **泛化能力**: 多任务学习提高泛化性能
-
-### 分类模式的优势
-1. **稳定性**: 冻结的backbone提供稳定的特征
-2. **效率**: 只训练分类头，训练速度快
-3. **防止过拟合**: 避免在小数据集上破坏预训练特征
-4. **专注性**: 专门针对分类任务优化
-
-## 🚀 实际使用建议
-
-1. **预训练阶段**: 使用全训练模式学习好的表示
-2. **微调阶段**: 使用分类模式，在预训练特征基础上优化分类性能
-3. **特征质量**: 分类模式依赖于全训练模式学到的高质量特征
-
-这种设计遵循了现代深度学习的"预训练-微调"范式，确保既能学到好的表示，又能在下游任务上获得最佳性能。
