--- PRETRAIN_FINETUNE_SEPARATION.md
@@ -1,112 +1,0 @@
-# SeqSetVAE: 预训练与微调完全分离设计
-
-## 🎯 设计理念
-
-根据你的要求，我们确保了预训练和微调阶段的**完全分离**：
-- **预训练阶段**: 维持原有的简单设计，专注于学习良好的数据表示
-- **微调阶段**: 应用现代VAE技术和优化策略，专注于分类性能
-
-## 📊 两个独立的模型类
-
-### 1. `SeqSetVAEPretrain` - 预训练模型
-**文件位置**: `model.py` 第130-503行
-
-**特点**:
-- ✅ **维持原有设计**: 完全保留你之前的预训练逻辑
-- ✅ **简单VAE特征**: 只使用后验均值 `mu`，不使用方差信息
-- ✅ **原始损失函数**: 重建损失 + KL损失，无分类损失
-- ✅ **无复杂特征融合**: 保持简单直接的特征提取
-
-**核心代码**:
-```python
-# 第363-365行: 简单的VAE特征提取
-_, z_list, _ = self.set_encoder(var, val)
-z_sample, mu, logvar = z_list[-1]
-z_prims.append(mu.squeeze(1))  # 只使用mu，保持原有设计
-```
-
-### 2. `SeqSetVAE` - 微调模型
-**文件位置**: `model.py` 第519-1583行
-
-**特点**:
-- 🚀 **现代VAE特征融合**: 同时使用mean和variance
-- 🚀 **完整预训练权重加载**: 智能加载兼容的预训练参数
-- 🚀 **完全参数冻结**: 除分类头外所有参数冻结
-- 🚀 **优化的分类性能**: 专门针对AUC/AUPRC优化
-
-**核心改进**:
-```python
-# 第982-993行: 现代VAE特征融合
-mu_feat = mu.squeeze(1)
-logvar_feat = logvar.squeeze(1)
-combined_feat = self._fuse_vae_features(mu_feat, logvar_feat)  # 使用mean+var
-z_prims.append(combined_feat)
-```
-
-## 🔄 训练流程分离
-
-### 预训练阶段
-```bash
-python train.py --mode pretrain \
-    --batch_size 8 \
-    --max_epochs 100
-```
-
-**使用的模型**: `SeqSetVAEPretrain`
-- 原有的简单设计
-- 只优化重建和KL损失
-- 学习良好的数据表示
-
-### 微调阶段
-```bash
-python train.py --mode finetune \
-    --pretrained_ckpt pretrain_checkpoint.ckpt \
-    --batch_size 4 \
-    --max_epochs 15
-```
-
-**使用的模型**: `SeqSetVAE`
-- 加载预训练权重
-- 冻结backbone，只训练分类头
-- 使用现代VAE特征融合
-
-## 📁 文件结构说明
-
-```
-model.py
-├── SetVAE (第51-128行)                    # 基础SetVAE模块
-├── SeqSetVAEPretrain (第130-503行)        # 预训练模型 - 原有设计
-└── SeqSetVAE (第519-1583行)               # 微调模型 - 改进设计
-    ├── _fuse_vae_features()               # 现代VAE特征融合
-    ├── _extract_enhanced_features()        # 增强特征提取
-    └── 完整的预训练权重加载逻辑
-```
-
-## 🔍 关键区别对比
-
-| 特性 | SeqSetVAEPretrain | SeqSetVAE |
-|------|------------------|-----------|
-| **VAE特征** | 只使用 `mu` | 使用 `mu + logvar` 融合 |
-| **权重加载** | 简单加载 | 智能兼容加载 |
-| **参数训练** | 全部参数训练 | 只训练分类头 |
-| **特征提取** | 基础pooling | 多尺度增强提取 |
-| **损失函数** | 重建 + KL | 分类 (+ 可选重建/KL) |
-| **目标** | 表示学习 | 分类性能优化 |
-
-## ✅ 确保分离的措施
-
-1. **代码隔离**: 两个完全独立的类，互不影响
-2. **训练脚本分离**: 根据mode选择不同的模型
-3. **配置分离**: 不同的超参数和优化策略
-4. **权重兼容性**: 智能映射确保预训练→微调的无缝衔接
-
-## 🚀 使用建议
-
-1. **先预训练**: 使用原有的`SeqSetVAEPretrain`学习表示
-2. **再微调**: 使用改进的`SeqSetVAE`优化分类性能
-3. **检查日志**: 确保模型类型和权重加载正确
-
-这种设计确保了：
-- ✅ 预训练阶段保持你原有的稳定设计
-- ✅ 微调阶段获得现代技术的性能提升
-- ✅ 两个阶段完全独立，互不干扰
