--- advanced_pretrain_proposal.py
@@ -1,150 +1,0 @@
-"""
-基于最新学术研究的VAE预训练改进方案
-Advanced VAE Pretraining based on latest research (2023-2024)
-"""
-
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-
-class UncertaintyAwarePretraining:
-    """
-    基于最新研究的不确定性感知预训练方案
-    参考：InfoVAE, Consistency Regularization, Uncertainty Quantification
-    """
-    
-    def __init__(self, model, alpha=0.1, beta=0.05):
-        self.model = model
-        self.alpha = alpha  # 不确定性权重
-        self.beta = beta   # 一致性权重
-    
-    def uncertainty_aware_feature_fusion(self, mu, logvar):
-        """
-        现代VAE特征融合：同时使用mean和variance
-        基于2024年最新研究的不确定性感知方法
-        """
-        # 方法1: 加权融合 (InfoVAE启发)
-        std = torch.exp(0.5 * logvar)
-        uncertainty_weight = torch.sigmoid(-std.mean(dim=-1, keepdim=True))
-        
-        # 方法2: 显式不确定性建模
-        uncertainty_features = std * 0.1  # 小权重避免不稳定
-        enhanced_mu = mu + uncertainty_features
-        
-        # 方法3: 自适应融合门控
-        fusion_gate = torch.sigmoid(mu + std)
-        fused_features = fusion_gate * mu + (1 - fusion_gate) * enhanced_mu
-        
-        return fused_features
-    
-    def consistency_regularization(self, mu1, logvar1, mu2, logvar2):
-        """
-        一致性正则化：确保编码器输出的一致性
-        基于 "Consistency Regularization for VAEs" 研究
-        """
-        # KL散度一致性
-        kl_consistency = F.kl_div(
-            F.log_softmax(mu1, dim=-1), 
-            F.softmax(mu2, dim=-1), 
-            reduction='batchmean'
-        )
-        
-        # 方差一致性
-        var_consistency = F.mse_loss(
-            torch.exp(0.5 * logvar1), 
-            torch.exp(0.5 * logvar2)
-        )
-        
-        return kl_consistency + var_consistency
-    
-    def enhanced_pretrain_forward(self, sets):
-        """
-        增强的预训练前向传播
-        结合最新的不确定性感知技术
-        """
-        z_prims, kl_total = [], 0.0
-        pos_list = []
-        
-        for i, s_dict in enumerate(sets):
-            var, val, time = s_dict["var"], s_dict["val"], s_dict["minute"]
-            
-            # 标准VAE编码
-            _, z_list, _ = self.model.set_encoder(var, val)
-            z_sample, mu, logvar = z_list[-1]
-            
-            # === 新增：不确定性感知特征融合 ===
-            if self.training:
-                # 训练时：使用不确定性感知融合
-                enhanced_features = self.uncertainty_aware_feature_fusion(mu, logvar)
-                z_prims.append(enhanced_features.squeeze(1))
-            else:
-                # 推理时：使用确定性特征（均值）
-                z_prims.append(mu.squeeze(1))
-            
-            # 标准KL损失计算
-            kl_div = 0.5 * torch.sum(torch.exp(logvar) + mu.pow(2) - 1 - logvar, dim=-1)
-            min_kl = self.model.free_bits * self.model.latent_dim
-            kl_div = torch.clamp(kl_div, min=min_kl)
-            
-            # === 新增：不确定性正则化 ===
-            uncertainty_reg = -self.alpha * torch.mean(logvar)  # 鼓励适度的不确定性
-            kl_total += kl_div.mean() + uncertainty_reg
-            
-            pos_list.append(time.unique().float())
-        
-        # 后续Transformer处理保持不变
-        kl_total = kl_total / len(sets)
-        z_seq = torch.stack(z_prims, dim=1)
-        
-        return z_seq, kl_total
-
-
-class ModernVAEPretrainStrategy:
-    """
-    现代VAE预训练策略的完整实现建议
-    """
-    
-    @staticmethod
-    def should_upgrade_pretraining(current_performance, stability_threshold=0.95):
-        """
-        评估是否应该升级预训练策略
-        """
-        criteria = {
-            "performance_stable": current_performance > stability_threshold,
-            "research_maturity": True,  # 2024年研究已相对成熟
-            "computational_cost": "moderate",  # 计算成本适中
-            "implementation_complexity": "low_to_moderate"
-        }
-        return criteria
-    
-    @staticmethod
-    def implementation_roadmap():
-        """
-        实施路线图建议
-        """
-        return {
-            "Phase 1": "保持现有预训练，优化微调阶段（已完成）",
-            "Phase 2": "A/B测试：对比传统vs不确定性感知预训练",
-            "Phase 3": "如果Phase 2效果好，逐步迁移到新方案",
-            "Phase 4": "长期：探索更先进的变分推理技术"
-        }
-
-
-# 使用示例
-def example_usage():
-    """
-    如何在你的代码中集成这些改进
-    """
-    # 在SeqSetVAEPretrain中添加选项
-    class EnhancedSeqSetVAEPretrain(SeqSetVAEPretrain):
-        def __init__(self, *args, use_uncertainty_aware=False, **kwargs):
-            super().__init__(*args, **kwargs)
-            self.use_uncertainty_aware = use_uncertainty_aware
-            if use_uncertainty_aware:
-                self.uncertainty_module = UncertaintyAwarePretraining(self)
-        
-        def _forward_single(self, sets):
-            if self.use_uncertainty_aware:
-                return self.uncertainty_module.enhanced_pretrain_forward(sets)
-            else:
-                return super()._forward_single(sets)  # 保持原有逻辑
