--- test_improvements.py
@@ -1,245 +1,0 @@
-#!/usr/bin/env python3
-"""
-Test script to verify the improvements made to SeqSetVAE
-验证SeqSetVAE改进的测试脚本
-"""
-
-import torch
-import torch.nn as nn
-from model import SeqSetVAE
-import config
-
-def test_pretrain_loading():
-    """测试预训练权重加载"""
-    print("🧪 Testing pretrained weight loading...")
-    
-    # 创建一个模拟的预训练checkpoint
-    mock_model = SeqSetVAE(
-        input_dim=config.input_dim,
-        reduced_dim=config.reduced_dim,
-        latent_dim=config.latent_dim,
-        levels=config.levels,
-        heads=config.heads,
-        m=config.m,
-        beta=config.beta,
-        lr=config.lr,
-        num_classes=config.num_classes,
-        ff_dim=config.ff_dim,
-        transformer_heads=config.transformer_heads,
-        transformer_layers=config.transformer_layers,
-        pretrained_ckpt=None,  # No pretrained for mock model
-    )
-    
-    # 保存mock model的状态
-    mock_state = mock_model.state_dict()
-    
-    # 创建新模型并尝试加载
-    test_model = SeqSetVAE(
-        input_dim=config.input_dim,
-        reduced_dim=config.reduced_dim,
-        latent_dim=config.latent_dim,
-        levels=config.levels,
-        heads=config.heads,
-        m=config.m,
-        beta=config.beta,
-        lr=config.lr,
-        num_classes=config.num_classes,
-        ff_dim=config.ff_dim,
-        transformer_heads=config.transformer_heads,
-        transformer_layers=config.transformer_layers,
-        pretrained_ckpt=None,  # Will test loading logic separately
-    )
-    
-    # 手动测试权重加载逻辑
-    print("   - Mock checkpoint created successfully")
-    print("   - Model architecture matches")
-    
-    return True
-
-def test_freeze_mechanism():
-    """测试参数冻结机制"""
-    print("🧪 Testing parameter freezing mechanism...")
-    
-    model = SeqSetVAE(
-        input_dim=config.input_dim,
-        reduced_dim=config.reduced_dim,
-        latent_dim=config.latent_dim,
-        levels=config.levels,
-        heads=config.heads,
-        m=config.m,
-        beta=config.beta,
-        lr=config.lr,
-        num_classes=config.num_classes,
-        ff_dim=config.ff_dim,
-        transformer_heads=config.transformer_heads,
-        transformer_layers=config.transformer_layers,
-        pretrained_ckpt=None,
-    )
-    
-    # 模拟finetune模式的冻结
-    model.enable_classification_only_mode()
-    
-    frozen_params = 0
-    trainable_params = 0
-    trainable_names = []
-    
-    for name, param in model.named_parameters():
-        if param.requires_grad:
-            trainable_params += param.numel()
-            trainable_names.append(name)
-        else:
-            frozen_params += param.numel()
-    
-    print(f"   - Frozen parameters: {frozen_params:,}")
-    print(f"   - Trainable parameters: {trainable_params:,}")
-    print(f"   - Trainable ratio: {trainable_params/(frozen_params+trainable_params)*100:.2f}%")
-    print(f"   - Trainable parameter names: {trainable_names}")
-    
-    # 验证只有分类头是可训练的
-    all_cls_head = all(name.startswith('cls_head') for name in trainable_names)
-    if all_cls_head and trainable_params > 0:
-        print("   ✅ Freezing mechanism working correctly")
-        return True
-    else:
-        print("   ❌ Freezing mechanism has issues")
-        return False
-
-def test_vae_feature_fusion():
-    """测试VAE特征融合"""
-    print("🧪 Testing VAE feature fusion...")
-    
-    model = SeqSetVAE(
-        input_dim=config.input_dim,
-        reduced_dim=config.reduced_dim,
-        latent_dim=config.latent_dim,
-        levels=config.levels,
-        heads=config.heads,
-        m=config.m,
-        beta=config.beta,
-        lr=config.lr,
-        num_classes=config.num_classes,
-        ff_dim=config.ff_dim,
-        transformer_heads=config.transformer_heads,
-        transformer_layers=config.transformer_layers,
-        pretrained_ckpt=None,
-    )
-    
-    # 测试VAE特征融合模块
-    batch_size = 2
-    latent_dim = config.latent_dim
-    
-    # 创建模拟的mu和logvar
-    mu = torch.randn(batch_size, latent_dim)
-    logvar = torch.randn(batch_size, latent_dim)
-    
-    # 测试两种模式
-    print("   - Testing full training mode...")
-    model.classification_only = False
-    fused_full = model._fuse_vae_features(mu, logvar)
-    print(f"     Full mode output shape: {fused_full.shape}")
-    
-    print("   - Testing classification-only mode...")
-    model.classification_only = True
-    fused_cls = model._fuse_vae_features(mu, logvar)
-    print(f"     Classification mode output shape: {fused_cls.shape}")
-    
-    # 验证输出形状正确
-    if fused_full.shape == (batch_size, latent_dim) and fused_cls.shape == (batch_size, latent_dim):
-        print("   ✅ VAE feature fusion working correctly")
-        return True
-    else:
-        print("   ❌ VAE feature fusion has shape issues")
-        return False
-
-def test_model_forward():
-    """测试模型前向传播"""
-    print("🧪 Testing model forward pass...")
-    
-    model = SeqSetVAE(
-        input_dim=config.input_dim,
-        reduced_dim=config.reduced_dim,
-        latent_dim=config.latent_dim,
-        levels=config.levels,
-        heads=config.heads,
-        m=config.m,
-        beta=config.beta,
-        lr=config.lr,
-        num_classes=config.num_classes,
-        ff_dim=config.ff_dim,
-        transformer_heads=config.transformer_heads,
-        transformer_layers=config.transformer_layers,
-        pretrained_ckpt=None,
-    )
-    
-    # 创建模拟输入
-    batch_size = 1
-    seq_len = 3
-    num_vars = 10
-    
-    # 创建模拟的患者序列数据
-    sets = []
-    for i in range(seq_len):
-        set_dict = {
-            'var': torch.randn(batch_size, num_vars, config.input_dim),
-            'val': torch.rand(batch_size, num_vars, 1),
-            'minute': torch.tensor([[i * 60.0]] * batch_size)  # 每小时一个set
-        }
-        sets.append(set_dict)
-    
-    try:
-        # 测试前向传播
-        model.eval()
-        with torch.no_grad():
-            logits, recon_loss, kl_loss = model(sets)
-        
-        print(f"   - Output logits shape: {logits.shape}")
-        print(f"   - Reconstruction loss: {recon_loss.item():.4f}")
-        print(f"   - KL loss: {kl_loss.item():.4f}")
-        print("   ✅ Model forward pass working correctly")
-        return True
-        
-    except Exception as e:
-        print(f"   ❌ Model forward pass failed: {e}")
-        return False
-
-def main():
-    """主测试函数"""
-    print("🚀 Testing SeqSetVAE Improvements")
-    print("=" * 50)
-    
-    tests = [
-        ("Pretrained Loading", test_pretrain_loading),
-        ("Freeze Mechanism", test_freeze_mechanism),
-        ("VAE Feature Fusion", test_vae_feature_fusion),
-        ("Model Forward", test_model_forward),
-    ]
-    
-    results = []
-    for test_name, test_func in tests:
-        print(f"\n📋 {test_name}")
-        try:
-            result = test_func()
-            results.append((test_name, result))
-        except Exception as e:
-            print(f"   ❌ Test failed with exception: {e}")
-            results.append((test_name, False))
-    
-    # 总结
-    print("\n" + "=" * 50)
-    print("📊 Test Results Summary:")
-    passed = 0
-    for test_name, result in results:
-        status = "✅ PASS" if result else "❌ FAIL"
-        print(f"   - {test_name}: {status}")
-        if result:
-            passed += 1
-    
-    print(f"\n🎯 Overall: {passed}/{len(results)} tests passed")
-    
-    if passed == len(results):
-        print("🎉 All improvements are working correctly!")
-    else:
-        print("⚠️ Some issues need to be addressed.")
-
-if __name__ == "__main__":
-    main()
